requirements.txt
tensorflow-cpu>=2.16
keras>=3.0
flask>=3.0
flask-cors>=4.0
pillow>=10.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
